requirements.txt
opencv-python==4.10.0.84
pillow==10.3.0
torch==2.5.1